ml-service/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
python-multipart==0.0.6
pefile==2023.2.7
numpy==1.26.3
requests==2.31.0
Pillow==10.2.0
# --- Test dependencies ---
pytest==8.0.0
pytest-cov==4.1.0
httpx==0.26.0
pandas==2.2.0
scikit-learn==1.4.0
joblib==1.3.2
slowapi==0.1.9

# --- OpenTelemetry ---
opentelemetry-api==1.22.0
opentelemetry-sdk==1.22.0
opentelemetry-instrumentation-fastapi==0.43b0
opentelemetry-exporter-otlp==1.22.0

# --- Prometheus ---
prometheus-fastapi-instrumentator==7.0.0

# --- Logging ---
structlog==24.1.0

